+++ src/grammar/GrammarConstants.js
@@ -7,22 +7,24 @@ GrammarConstants.catchAllKeyword = "@catchAllKeyword"
 GrammarConstants.defaults = "@defaults"
 GrammarConstants.constants = "@constants"
 
-// executing
-GrammarConstants.parseClass = "@parseClass"
+// virtual machines instantiating and executing
+GrammarConstants.parseClass = "@parseClass" // todo: make a map to support different cc and vms
+// @machines js
 
 // compiling
-GrammarConstants.compiled = "@compiled"
-GrammarConstants.compiledIndentCharacter = "@compiledIndentCharacter"
-GrammarConstants.listDelimiter = "@listDelimiter"
-GrammarConstants.openChildren = "@openChildren"
-GrammarConstants.closeChildren = "@closeChildren"
-GrammarConstants.targetExtension = "@targetExtension"
+GrammarConstants.compilerKeyword = "@compiler"
+GrammarConstants.compiler = {}
+GrammarConstants.compiler.sub = "@sub" // replacement instructions
+GrammarConstants.compiler.indentCharacter = "@indentCharacter"
+GrammarConstants.compiler.listDelimiter = "@listDelimiter"
+GrammarConstants.compiler.openChildren = "@openChildren"
+GrammarConstants.compiler.closeChildren = "@closeChildren"
 
 // developing
 GrammarConstants.description = "@description"
 GrammarConstants.frequency = "@frequency"
 
-// ohayo ui
+// ohayo ui. todo: remove all.
 GrammarConstants.ohayoSvg = "@ohayoSvg"
 GrammarConstants.ohayoTileSize = "@ohayoTileSize"
 GrammarConstants.ohayoTileClass = "@ohayoTileClass"