+++ src/grammar/TreeGrammarProgram.js
@@ -1,10 +1,13 @@
 const TreeProgram = require("../TreeProgram.js")
 const fs = require("fs")
-const TreeGrammar = fs.readFileSync(__dirname + "/TreeGrammar.grammar", "utf8")
 
 class TreeGrammarProgram extends TreeProgram {
   getGrammarString() {
-    return TreeGrammar
+    return fs.readFileSync(__dirname + "/TreeGrammar.grammar", "utf8")
+  }
+
+  getGrammarFilePath() {
+    return __dirname + "/TreeGrammar.grammar"
   }
 }